requirements.txt
asgiref==3.4.1
attrs==22.2.0
backports.zoneinfo==0.2.1
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==40.0.2
defusedxml==0.7.1
dj-rest-auth==5.1.0
Django==3.2.25
django-allauth==0.53.1
django-cors-headers==3.10.1
django-environ==0.11.2
django-rest-auth==0.9.5
djangorestframework==3.15.1
djangorestframework-simplejwt==4.4.0
drf-spectacular==0.27.1
idna==3.7
importlib-metadata==4.8.3
importlib-resources==5.4.0
inflection==0.5.1
jsonschema==4.0.0
oauthlib==3.2.2
Pillow==8.4.0
pycparser==2.21
PyJWT==2.4.0
pyrsistent==0.18.0
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
requests==2.27.1
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.18
zipp==3.6.0
